Create JELDO MONEY App is a command-line tool that generates a fully configured starter JELDO MONEY (GraphQL) project with only essential dependencies. With Create JELDO MONEY App, you can start developing a client, server, or mobile app, or a universal application in no time.
The projects generated with Create JELDO MONEY App are built of familiar technologies from the JavaScript ecosystem:
TypeScript, Webpack, React, Express, and React Native. Depending on the type of the project you want to develop, the
mentioned technologies are used in different combinations to provide particular setups for server, client, or mobile
development.